# docs/swagger.yaml
basePath: /
definitions:
  models.Client:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      patronymic:
        type: string
      phone_number:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample serice for managing clients
  termsOfService: http://swagger.io/terms/
  title: Clients API
  version: "1.0"
paths:
  /clients:
    get:
      consumes:
      - application/json
      description: Get  all clients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
      summary: Get  all clients
      tags:
      - Clients
    post:
      consumes:
      - application/json
      description: Create a new client with the input paylod
      parameters:
      - description: Create client
        in: body
        name: newClient
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
      summary: Create a new client
      tags:
      - Clients
  /clients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete particular client by id
      parameters:
      - description: Delete Client
        in: body
        name: idToDelete
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      summary: Delete particular client
      tags:
      - Clients
    get:
      description: get client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      summary: Get one client
      tags:
      - Clients
    put:
      consumes:
      - application/json
      description: Update particular client by id
      parameters:
      - description: Update Client
        in: body
        name: updatedclient
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      summary: Update particular client
      tags:
      - Clients
swagger: "2.0"